refactor: move to composable
