+++ components/items/ItemsGrid/ItemsGridImage.vue
@@ -1,18 +1,17 @@
 <template>
   <NeoNftCard
-    v-if="item"
-    :nft="item"
+    v-if="nft"
+    :nft="nft"
     :prefix="urlPrefix"
-    :show-price="Number(item?.price) > 0"
+    :show-price="Number(nft?.price) > 0"
     :variant="variant" />
 </template>
 
 <script setup lang="ts">
 import { NeoNftCard } from '@kodadot1/brick'
-import { getNftMetadata } from '@/utils/nft'
 
 import type { NftCardVariant } from '@kodadot1/brick'
-import type { NFTWithMetadata } from '@/utils/nft'
+import type { NFTWithMetadata } from '@/composables/useNft'
 
 const { urlPrefix } = usePrefix()
 
@@ -21,9 +20,5 @@ const props = defineProps<{
   variant?: NftCardVariant
 }>()
 
-const item = ref<NFTWithMetadata>()
-
-onBeforeMount(async () => {
-  item.value = await getNftMetadata(props.nft, urlPrefix.value)
-})
+const { nft } = useNft(props.nft)
 </script>
